bitwiseinstruction

+++ src/control.h
@@ -4,10 +4,11 @@
 
 #ifndef ARMV8_25_CONTROL_H
 #define ARMV8_25_CONTROL_H
+#include <stdint.h>
 
 typedef enum  {IMMEDIATE, REGISTER, LOADSTORE, BRANCH, FIRST = IMMEDIATE, LAST = BRANCH, UNDEFINED} instructionType;
 
-int getBits(int start, int end, int instruction);
+int64_t getBits(int start, int end, int64_t instruction);
 
 instructionType getInstructionType(int instruction);
